feat(simu): split queuing audio from SDL
