+++ radio/src/targets/simu/CMakeLists.txt
@@ -5,13 +5,12 @@ if(NOT SIMU_TARGET)
   return()
 endif()
 
-target_compile_options(radiolib_native PUBLIC -DSIMU)
-
 set(SIMU_DRIVERS
   simpgmspace.cpp
   simufatfs.cpp
   simudisk.cpp
   simulcd.cpp
+  audio_driver.cpp
   switch_driver.cpp
   adc_driver.cpp
   module_drivers.cpp
@@ -40,7 +39,6 @@ include_directories(${CMAKE_CURRENT_BINARY_DIR})
 if(SDL2_FOUND)
   include_directories(${SDL2_INCLUDE_DIRS})
   target_compile_options(radiolib_native PUBLIC -DSIMU_AUDIO)
-  list(APPEND SIMU_DRIVERS simuaudio.cpp)
   if(TARGET SDL2::SDL2)
     set(SDL2_LIBRARIES SDL2::SDL2)
   endif()
@@ -97,20 +95,24 @@ if(Qt6Widgets_FOUND)
   qt_wrap_cpp(SIMULATOR_SRC
     ${COMPANION_SRC_DIRECTORY}/simulation/simulatorinterface.h
     opentxsimulator.h
-    )
+  )
   list(APPEND SIMULATOR_SRC ${SIMU_SRC} opentxsimulator.cpp)
+
   add_library(${SIMULATOR_TARGET} SHARED ${SIMULATOR_SRC})
+  if(SDL2_FOUND)
+    target_sources(${SIMULATOR_TARGET} PRIVATE simuaudio.cpp)
+  else()
+    target_sources(${SIMULATOR_TARGET} PRIVATE no_audio.cpp)
+  endif()
 
   target_compile_options(${SIMULATOR_TARGET} PRIVATE ${SIMU_SRC_OPTIONS})
   target_compile_definitions(${SIMULATOR_TARGET} PUBLIC ${APP_COMMON_DEFINES})
   target_link_libraries(${SIMULATOR_TARGET} ${SDL2_LIBRARIES} Qt::Core Qt::SerialPort)
 
   # Remove debug symbols on release builds
-  if(NOT CLANG)
-    if("${CMAKE_BUILD_TYPE}" STREQUAL "Release")
-      add_custom_command(TARGET ${SIMULATOR_TARGET} POST_BUILD
-        COMMAND ${CMAKE_STRIP} $<TARGET_FILE:${SIMULATOR_TARGET}>)
-    endif()
+  if("${CMAKE_BUILD_TYPE}" STREQUAL "Release")
+    add_custom_command(TARGET ${SIMULATOR_TARGET} POST_BUILD
+      COMMAND ${CMAKE_STRIP} -x $<TARGET_FILE:${SIMULATOR_TARGET}>)
   endif()
 
   add_custom_command(
@@ -118,7 +120,7 @@ if(Qt6Widgets_FOUND)
     COMMAND ${CMAKE_COMMAND} -E make_directory ${CMAKE_BINARY_DIR}/plugins/
     COMMAND ${CMAKE_COMMAND} -E copy
             $<TARGET_FILE:${SIMULATOR_TARGET}> ${CMAKE_BINARY_DIR}/plugins/
-    )
+  )
 
   # Introduced at 2.5 - TODO replace with a code change
   # When compiled shared simulators using libopenui and Companion have classes named MainWindow
@@ -171,6 +173,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${WARNING_FLAGS}")
 add_executable(simu
   EXCLUDE_FROM_ALL
   ${SIMU_SRC}
+  simuaudio.cpp
   display.cpp
   widgets.cpp
   knobs.cpp
