+++ radio/src/tests/gtests.cpp
@@ -127,6 +127,8 @@ uint16_t simu_get_analog(uint8_t idx)
   return 0;
 }
 
+void simuQueueAudio(const uint8_t*, uint32_t) {}
+
 static char _stringResult[200];
 
 const char * nchar2string(const char * string, int size)
